pyproject.toml
[tool.poetry]
name = "laser_serial_reader"
version = "0.1.0"
description = "Read from the serial monitor and do things"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
pygame = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
